Add throttle network on the frame session
A page can be made up of many frames, and each frame has its own
session. This session maps directly to one in chrome. Generally a
website will have one main frame.

The new throttling api is scoped to the page so all the frames will
need to have their sessions updated for a single page. This change
helps with that.