Fix isHidden by using runActionOnSelector
isHidden will now work with runActionOnSelector just like isVisible.
This means that isHidden will not wait for the selector to match on an
element before returning the result. It doesn't make sense for the API
to wait for an element to match before checking whether it is hidden
or not.

Note that when no matches are found for the selector, it return true
instead of false, which is to say that the element is hidden (or
hasn't been found).

This makes timeout obsolete for isHidden too, just like it is for
isVisible.