Add parse console values
These new functions are to be used when parsing the console message
which is being printed by the website under test when it used
console.log. Since these are string values that will be parsed and
reprinted to stdout, we don't necessarily need to parse a json string
into a go object (which is the main motivation for this change). This
will avoid situations where k6 browser panics and causes the test run
to end due to not being able to parse a json string.

Other areas of the implementation (e.g. toJSON or eval) still require
a go representation of the object and they will work with the existing
parse functions.