Refactor TraceNavigation
Based on discussions in #1140 , a decision was made to, instead of using
the URL as the span name for navigation spans, use the word 'navigation'
to identify the action, following the same pattern as the rest of the
spans. Due to this, and because a navigation span will always have that
fixed span name, it is no longer necessary to accept the span name as
input for the method.