Refactor elementHandle.click parsing
Move the parsing of the options for elementHandle.click outside of the
promise goroutine and back onto the main goja thread in the mapping
layer. This will help mitigate the risk of a panic if more than one
goroutine is accessing the goja runtime off the main goja thread.

This doesn't solve the problem completely though since this API calls
to other areas of the codebase which does still interact with the goja
runtime. See https://github.com/grafana/xk6-browser/issues/1170 for
further details.