Refactor parse out of frame.click
This helps avoid using the goja runtime when it comes to parsing the
options. The options parsing has been moved to the mapping layer and so
has been moved back on to the main goja thread, which will mitigate the
possibility of a panic which could occur when multiple goroutines try
to work with the goja runtime (which is not thread safe).

There is another issue to tackle more of the goja refactoring in
https://github.com/grafana/xk6-browser/issues/1170.

The change also adds a temporary fix for page.click which will also be
refactored at a later stage.