Refactor parse in frame.waitForFunction
Refactor the parsing of the optional options out of the
frame.waitForFunction promise, and move it into the mapping layer which
is on the main goja thread and out of the promise. This will mitigate
the risk of accessing the non thread safe goja runtime from more than
one goroutine.

This change also adds a temporary change in page.waitForFunction to
parse the goja value, which will also change at a later stage.