Move goja.Value args transform
This moves the transformation of the args for the frame.waitForFunction
API to the mapping layer and out of the promise. This will help
mitigate the risk of using the non thread safe goja runtime from
multiple goroutines, and only work with it from the main goja thread in
the mapping layer.

This adds a temporary change to page.waitForFunction where it still
does the transform within the promise. This will be moved at a later
stage.