Add startIteration helper
This function will be used by TestPageWaitForFunction. That test
requires access to the mapped versions of page. Instead of exporting a
new MapPage method in mapping.go, work with the existing method to
create a new module instance, set the mapped browser to the runtime
and work with the event system.