Refactor to work with the filePersister
We're refactoring out the existing use of working directly with the os
package to persist screenshots to the local disk from screenshotter.
This change enables to work with the localPersister whereby the
underlying details of the where and how are hidden allowing us to
easily use either the localPersister (current) or the remotePersister
(future).