Move newFilePersister to browser package
The browser package is the work horse on initialising and setting up
the environment for the global test run, iteration or vu. The decision
on whether to create a Local or Remote persister should rest in the
browser package's hands, but leave the details of what and how the two
persisters do their job in the storage package.