Add Connection.onTargetAttachedToTarget callback
This callback will be called to manage the connection session handling
from outside. Since Browser is the creator of Connection, we're adding
this callback to let Browser hook into target attachment and tell
Connection to stop attaching the target if Browser doesn't want it.

This is a downside of the current design. There is a symbiotic
relationship between Connection and Browser. Perhaps they should be
together, not separate this way. But for the current design, this is the
best we can do to improve Browser's session management.