Fix error handling from chromium
When the browser module receives the error with code -32000, this
indicates a server error. From looking at the chromium code it seems to
be related to issues that originate not from chromium but from the
caller. There are many error messages that use this error code, so we
cannot assume that it is due to a navigation. The user and we will
benefit from knowing what chrome/chromium is not happy about

Resolves: https://github.com/grafana/xk6-browser/pull/1233#discussion_r1512469217