Fix iframe execCtx race
The events are as follows when an iframe is attached:
1. execCtx for about:blank for new frame is set;
2. execCtx for navigated frame with url is set;
3. original about:blank execCtx is destroyed.

If we only work with the original execCtx, then it will be destroyed by
the 3rd event. To prevent this we overwrite the original one with the
new one in the 2nd event.

What this means is that we do not end up in an infinite loop when
waitForExecutionContext is called.