Add fix to prevent panic when ctx closed
If the context is closed, and that is the error that is received by the
dispose method, then that would elude to the fact that the iteration
has ended. Therefore, we shouldn't need to panic or log an error at
this point, and instead just continue on as normal. The page has closed
well before this panic occurs.