Add error handling for Dispose
Dispose can panic when a navigation occurs on a page with an iframe.
The specific error that is returned is "Cannot find context with
specified id". The reason for this error is due to the fact that Chrome
has already cleaned up the associated resource before it receives the
CDP command to release it. Therefore it is safe to handle this error,
log a debug log and carry on with the test.