Refactor RunIfWaitingForDebugger
We have identified that the call to RunIfWaitingForDebugger wasn't
being performed at the correct time when a new page was created after a
link was clicked on that opened the page.

The first step is to move out of the initOptions so that we can
eventually run it independently of the other setup/init steps.