Add fix to prevent iframe panic on navigation
The change prevents the iframe child frames from being removed when a
detach event comes in from Chrome.

An iframe seems to be associated to multiple sessions. The initial
sessions that is linked to the mainframe that the iframe belongs to and
its own session. When the page navigates, a detach event arrives with
the initial session and a swap type for the iframe. The iframe though
is associated to its own session. In this scenario we want to avoid
removing frames, which prevents panics occurring when the iframe
navigates.