Move handleFrameTree
handleFrameTree is called when a new frame session is created for a new
frame that is attached. Some of these frames are not mainframes but do
contains a tree structures with child frames. If we don't handle the
frame tree for these non-mainframes, this will prevent navigations from
being handled for these frames and a panic/error which would stop the
test iteration.