Improve null detection in parseRemoteObjectValue
Previously, we can't detect a null value:

        v, _ := page.GetAttribute("#el", "missing", nil)
        assert.Nil(t, nil, v) // v is not nil

Now, `v` will be `nil` instead of "null" as a string. This enables us to
detect `nil` in methods like GetAttribute, TextNode, etc. to tell the
user that what they're looking for does not exist instead of returning
"null" or an empty string.