Rename ctx to backgroundCtx
This helps us by clarifying that this is a background context, and not
controlled by k6 (i.e. the VU context). The background context is to
only be used to control the lifecycle of chromium/connection. In later
commits vu context will be reintroduced to control the iteration
lifecycle so that k6 can abort the iteration and the browser module
doesn't block that from happening.