Use two contexts
The background context will be used for the chromium subprocess,
connection and things that need to be controlled by the iterStart and
iterEnd events. The k6Ctx is the context that is controlled by the vu
context. This will allow the iteration to abort due to a SIGTERM while
also allowing chromium to exit with endIter/exit events. This seems to
avoid any race conditions.