Use two contexts
The background context will be used for the chromium subprocess,
connection and things that need to be controlled by the iterStart and
iterEnd (k6 event system) events. The k6Ctx is the context that is
controlled by the vu context. This will allow the iteration to abort
due to a SIGTERM while also allowing chromium to exit with endIter/exit
events.

At the moment this only introduces the k6Ctx, but it is still the
background context.