Add skip on chromium crash tests
When an invalid value is used for --remote-debugging-port chromium used
to return an error. Now it just hangs. We've opened an issue with the
chromium team to understand whether this is a regression that will be
fixed or whether we need to change/delete these tests.

The solution to unblock us for now is to skip these tests.

https://issues.chromium.org/issues/364089353