Fix the navigation span's starting point
The navigation span was started a lot later than it should have. This
meant that child spans (such as web vitals) looked incorrect since
the values of those child spans were longer than the navigation span.

Now the navigation span is started a lot earlier, and this means that
the child span values are safely within the permitted navigation span.

We still have a navigation span start from onFrameNavigated which is
needed when the iteration starts on a new page, since
onFrameStartedLoading isn't called when a new page is called and it
navigates to about:blank.