Add context cancel func in connection
The context needs to be closed when the connection is closed. We're
using the context to ensure that this happens. The reason we need to
close the context is so that the BaseEventEmitter is correctly shutdown
when the connection is closed.

This is being added here incase the context coming into NewConnection
cannot be cancelled.