Refactor runtime out of vy context in tests
A recent change was added to k6 engine, and I soon realised that there
was something broken with the vu context. The vu context doesn't
automatically come with the runtime, this is added by us.

I wanted to ensure that we weren't surprised when adding the latest
change in the k6 engine, so i've removed the runtime when the vu
context is created for the test.

This means that if we make a change in the codebase that affects the
runtime in the vu context we will find out before any changes are added
in the k6 engine.