Fix k6 process hang when using page.on
This fix ensures that the taskqueue is closed when the iteration ends.
If the taskqueue isn't closed when the iteration ends the k6 process
will hang indefinitely. We used to rely on page.close, but now that the
life cycle of the iteration is controlled by the VU we can safely rely
on the VU context to ensure that the taskqueue is closed (if it isn't
already closed).