Add synchronization and name propagation
The name needs to be sent back to the caller that needs it. It is done
via a channel.

When running things on the taskqueue, they are done async. We need to
make this process synchronous. We're relying on the nameCh (the same
channel that the name is sent to the caller on) to notify the called
when the handler has complete.