Refactor how the name is returned back to caller
Instead of working with goroutines and channels to propagate the name
back from GroupURLTag, a name field is updated in ExportedMetric. The
call to the taskqueue is synchronised so when the handler is called we
can safely wait for it to complete instead of relying on channels and
contexts to wait for it to complete.