Fix recvloop so it doens't use as much memory
recvloop was using ws.ReadMessage, which was in turn working with
io.ReadAll. It turns out for the current use, this isn't a good way to
read the message of the web socket connection. Instead we should buffer
the read, store it in a buffer that can increase with the test, which
is then cloned before being unmarshaled. We have more control over the
memory by working in this way.