Fix memory leak in queue
When we read off the read queue, we alias the slice. What has been
forgotten though is to remove the reference to the entry in the slice
after it has been read. This fixes that by overwriting the entry with a
new Event. This now allows the GC to pick up the buffered data that the
event was holding onto.